chore: extracted Stream Store acceptance tests
This will allow us to use the same test suite for new implementations.
