+++ src/postgres/__tests__/pg-stream-store.pg-notifications.test.ts
@@ -2,7 +2,7 @@ import { v4 } from 'uuid'
 import { createPostgresStreamStore, PgStreamStore } from '../pg-stream-store'
 import { streamStoreCfg } from '../../__helpers__/pg-stream-store-config'
 import { ExpectedVersion } from '../../types/stream-store'
-import { generateMessages } from './__helpers__/message-helper'
+import { generateMessages } from '../../__helpers__/message-helper'
 import { SubscribeAt } from '../../types/subscriptions'
 import { delay } from '../../utils/promise-util'
 import { noopLogger } from '../../logging/noop'
